Implement script for calculating missing bumblebee checksums as browser view:
This is needed because for some reason, when invoked via bin/instance0 run,
collective.taskqueue chokes with
"AssertionError: No task queues defined. Cannot queue task."
